# swagger.yaml
openapi: 3.0.0
info:
  title: The System MySQL Database
  description: The System MySQL Database
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /contacts:
    get:
      summary: Get a list of all contacts
      description: Retrieve list of all contacts.
      operationId: getAllContacts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 123
                    primaryContactName:
                      type: string
                      example: 'Mickey Mouse'
                    primaryContactNumber:
                      type: string
                      example: '0123456789'
                    primaryContactEmail:
                      type: string
                      example: 'mickey@example.com'
                    representativeName:
                      type: string
                      example: 'Donald Duck'
                    representativeNumber:
                      type: string
                      example: '0123456789'
                    representativeEmail:
                      type: string
                      example: 'donald@example.com'
                    alternativeRepName:
                      type: string
                      example: 'Goofy'
                    alternativeRepNumber:
                      type: string
                      example: '0123456789'
                    alternativeRepEmail:
                      type: string
                      example: 'goofy@example.com'
                    otherNumber1:
                      type: string
                      example: '0123456789'
                    otherNumber2:
                      type: string
                      example: '0123456789'
                    otherNumber3:
                      type: string
                      example: '0123456789'
                    otherNumber4:
                      type: string
                      example: '0123456789'
                    otherNumber5:
                      type: string
                      example: '0123456789'
                    otherNumber6:
                      type: string
                      example: '0123456789'
                    otherNumber7:
                      type: string
                      example: '0123456789'
                    otherNumber8:
                      type: string
                      example: '0123456789'
                    otherNumber9:
                      type: string
                      example: '0123456789'
                    otherNumber10:
                      type: string
                      example: '0123456789'
                    otherEmail1:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail2:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail3:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail4:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail5:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail6:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail7:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail8:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail9:
                      type: string
                      example: 'minnie@example.com'
                    otherEmail10:
                      type: string
                      example: 'minnie@example.com'
                    dnc1:
                      type: string
                      example: '0123456789'
                    dnc2:
                      type: string
                      example: '0123456789'
                    dnc3:
                      type: string
                      example: '0123456789'
                    dnc4:
                      type: string
                      example: '0123456789'
                    dnc5:
                      type: string
                      example: '0123456789'
                    createdDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    createdBy:
                      type: string
                      example: 'mickey@example.com'
                    updatedDate:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
                    updatedBy:
                      type: string
                      example: 'mickey@example.com'
                    f_accountNumber:
                      type: string
                      example: 'AAA101'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    post:
      summary: Create a new contact
      description: Restricted to admin users.
      operationId: createContact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  primaryContactName:
                    type: string
                    example: 'Mickey Mouse'
                  primaryContactNumber:
                    type: string
                    example: '0123456789'
                  primaryContactEmail:
                    type: string
                    example: 'mickey@example.com'
                  representativeName:
                    type: string
                    example: 'Donald Duck'
                  representativeNumber:
                    type: string
                    example: '0123456789'
                  representativeEmail:
                    type: string
                    example: 'donald@example.com'
                  alternativeRepName:
                    type: string
                    example: 'Goofy'
                  alternativeRepNumber:
                    type: string
                    example: '0123456789'
                  alternativeRepEmail:
                    type: string
                    example: 'goofy@example.com'
                  otherNumber1:
                    type: string
                    example: '0123456789'
                  otherNumber2:
                    type: string
                    example: '0123456789'
                  otherNumber3:
                    type: string
                    example: '0123456789'
                  otherNumber4:
                    type: string
                    example: '0123456789'
                  otherNumber5:
                    type: string
                    example: '0123456789'
                  otherNumber6:
                    type: string
                    example: '0123456789'
                  otherNumber7:
                    type: string
                    example: '0123456789'
                  otherNumber8:
                    type: string
                    example: '0123456789'
                  otherNumber9:
                    type: string
                    example: '0123456789'
                  otherNumber10:
                    type: string
                    example: '0123456789'
                  otherEmail1:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail2:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail3:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail4:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail5:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail6:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail7:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail8:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail9:
                    type: string
                    example: 'minnie@example.com'
                  otherEmail10:
                    type: string
                    example: 'minnie@example.com'
                  dnc1:
                    type: string
                    example: '0123456789'
                  dnc2:
                    type: string
                    example: '0123456789'
                  dnc3:
                    type: string
                    example: '0123456789'
                  dnc4:
                    type: string
                    example: '0123456789'
                  dnc5:
                    type: string
                    example: '0123456789'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_accountNumber:
                    type: string
                    example: 'AAA101'
              required:
                - f_accountNumber
      responses:
        '200':
          description: Contact created successfully. The id of the new contact is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
        '400':
          description: Contact id already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Contact id '123' already exists"
        '401':
          $ref: '#/components/responses/UnauthorisedError'
...